requirements.txt
streamlit
requests
arxiv
python-dotenv
fpdf2
groq